add lofire-net crate
