+++ lofire-demo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lofire-app"
+name = "lofire-demo"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,8 +7,6 @@ edition = "2021"
 
 [dependencies]
 lofire = { path = "../lofire" }
+lofire-net = { path = "../lofire-net" }
 lofire-broker = { path = "../lofire-broker" }
-
-[dependencies.async-std]
-version = "1.7.0"
-features = ["attributes"]
+async-std = {  version = "1.7.0", features = ["attributes"] }
