+++ lofire-net/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lofire-p2p"
+name = "lofire-net"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,9 +7,6 @@ edition = "2021"
 
 [dependencies]
 lofire = { path = "../lofire" }
-lofire-net = { path = "../lofire-net" }
-blake3 = "1.3.1"
-chacha20 = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_bare = "0.5.0"
 serde_bytes = "0.11.7"
