+++ lofire-net/src/types.rs
@@ -1,10 +1,561 @@
-//! LoFiRe broker protocol types
+//! LoFiRe network protocol types
 //!
 //! Corresponds to the BARE schema
 
 use lofire::types::*;
-use lofire_p2p::types::*;
 use serde::{Deserialize, Serialize};
+
+//
+// COMMON TYPES FOR MESSAGES
+//
+
+/// Peer ID: public key of the node
+pub type PeerId = PubKey;
+
+/// Overlay ID
+///
+/// - for public overlays that need to be discovered by public key:
+///   BLAKE3 hash over the repository public key
+/// - for private overlays:
+///   BLAKE3 keyed hash over the repository public key
+///   - key: BLAKE3 derive_key ("LoFiRe OverlayId BLAKE3 key", repo_secret)
+pub type OverlayId = Digest;
+
+/// Overlay session ID
+///
+/// Used as a component for key derivation.
+/// Each peer generates it randomly when (re)joining the overlay network.
+pub type SessionId = u64;
+
+/// Topic ID: public key of the topic
+pub type TopicId = PubKey;
+
+/// Result codes
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub enum Result {
+    Ok,
+    Error,
+}
+
+/// IPv4 address
+pub type IPv4 = [u8; 4];
+
+/// IPv6 address
+pub type IPv6 = [u8; 16];
+
+/// IP address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub enum IP {
+    IPv4(IPv4),
+    IPv6(IPv6),
+}
+
+/// IP transport protocol
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub enum IPTransportProtocol {
+    TLS,
+    QUIC,
+}
+
+/// IP transport address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub struct IPTransportAddr {
+    pub ip: IP,
+    pub port: u16,
+    pub protocol: IPTransportProtocol,
+}
+
+/// Network address
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub enum NetAddr {
+    IPTransport(IPTransportAddr),
+}
+
+//
+// OVERLAY MESSAGES
+//
+
+/// Overlay connection request
+///
+/// Sent to an existing overlay member to initiate a session
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum OverlayConnect {
+    V0(),
+}
+
+/// Overlay disconnection request
+///
+/// Sent to a connected overlay member to terminate a session
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum OverlayDisconnect {
+    V0(),
+}
+
+/// Content of TopicAdvertV0
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct TopicAdvertContentV0 {
+    /// Topic public key
+    pub topic: TopicId,
+
+    /// Peer public key
+    pub peer: PeerId,
+}
+
+/// Topic advertisement by a publisher
+///
+/// Flooded to all peers in overlay
+/// Creates subscription routing table entries
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct TopicAdvertV0 {
+    pub content: TopicAdvertContentV0,
+
+    /// Signature over content by topic key
+    pub sig: Sig,
+}
+
+/// Topic advertisement by a publisher
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum TopicAdvert {
+    V0(TopicAdvertV0),
+}
+
+/// Topic subscription request by a peer
+///
+/// Forwarded towards all publishers along subscription routing table entries
+/// that are created by TopicAdverts
+/// Creates event routing table entries along the path
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct SubReqV0 {
+    /// Random ID generated by the subscriber
+    pub id: u64,
+
+    /// Topic public key
+    pub topic: TopicId,
+}
+
+/// Topic subscription request by a peer
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum SubReq {
+    V0(SubReqV0),
+}
+
+/// Topic subscription acknowledgement by a publisher
+///
+/// Sent to all subscribers in an Event.
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct SubAckV0 {
+    /// SubReq ID to acknowledge
+    pub id: u64,
+}
+
+/// Topic subscription acknowledgement by a publisher
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum SubAck {
+    V0(SubAckV0),
+}
+
+/// Topic unsubscription request by a subscriber
+///
+/// A broker unsubscribes from upstream brokers
+/// when it has no more subscribers left
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct UnsubReqV0 {
+    /// Topic public key
+    pub topic: TopicId,
+}
+
+/// Topic unsubscription request by a subscriber
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum UnsubReq {
+    V0(UnsubReqV0),
+}
+
+/// Topic unsubscription acknowledgement
+/// Sent to the requestor in response to an UnsubReq
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct UnsubAckV0 {
+    /// Topic public key
+    pub topic: TopicId,
+}
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum UnsubAck {
+    V0(UnsubAckV0),
+}
+
+/// Branch change notification
+/// Contains a chunk of a newly added Commit or File referenced by a commit.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ChangeV0 {
+    /// Block with encrypted content
+    pub content: Block,
+
+    /// Encrypted key for the Commit object in content
+    /// Only set for the root block of the object
+    /// The key is encrypted using ChaCha20:
+    /// - key: BLAKE3 derive_key ("LoFiRe Event ObjectRef ChaCha20 key",
+    ///                           branch_pubkey + branch_secret + publisher_pubkey)
+    /// - nonce: commit_seq
+    pub key: Option<SymKey>,
+}
+
+/// Body of EventContentV0
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum EventBodyV0 {
+    SubAck,
+    Change,
+}
+
+/// Content of EventV0
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct EventContentV0 {
+    /// Pub/sub topic
+    pub topic: TopicId,
+
+    /// Publisher pubkey hash
+    /// BLAKE3 keyed hash over branch member pubkey
+    /// - key: BLAKE3 derive_key ("LoFiRe Event publisher BLAKE3 key",
+    ///                           repo_pubkey + repo_secret +
+    ///                           branch_pubkey + branch_secret)
+    pub publisher: Digest,
+
+    /// Commit sequence number of publisher
+    pub seq: u32,
+
+    /// Event body
+    pub body: EventBodyV0,
+}
+
+/// Pub/sub event published in a topic
+///
+/// Forwarded along event routing table entries
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct EventV0 {
+    pub content: EventContentV0,
+
+    /// Signature over content by topic key
+    pub sig: Sig,
+}
+
+/// Pub/sub event published in a topic
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub enum Event {
+    V0(EventV0),
+}
+
+/// Object search in a pub/sub topic
+///
+/// Sent along the reverse path of a pub/sub topic
+/// from a subscriber to all publishers.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BlockSearchTopicV0 {
+    /// Topic to forward the request in
+    pub topic: TopicId,
+
+    /// List of Object IDs to request
+    pub ids: Vec<ObjectId>,
+
+    /// Whether or not to include all children recursively in the response
+    pub include_children: bool,
+
+    /// List of Peer IDs the request traversed so far
+    pub path: Vec<PeerId>,
+}
+
+/// Object request by ID
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum BlockSearchTopic {
+    V0(BlockSearchTopicV0),
+}
+
+/// Block search along a random walk
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BlockSearchRandomV0 {
+    /// List of Block IDs to request
+    pub ids: Vec<BlockId>,
+
+    /// Whether or not to include all children recursively in the response
+    pub include_children: bool,
+
+    /// Number of random nodes to forward the request to at each step
+    pub fanout: u8,
+
+    /// List of Peer IDs the request traversed so far
+    pub path: Vec<PeerId>,
+}
+
+/// Block request by ID using a random walk
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum BlockSearchRandom {
+    V0(BlockSearchRandomV0),
+}
+
+/// Response to a BlockSearch* request
+///
+/// Follows request path with possible shortcuts.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BlockResultV0 {
+    /// Response path
+    pub path: Vec<PeerId>,
+
+    /// Resulting Object(s)
+    pub payload: Vec<Block>,
+}
+
+/// Response to a BlockSearch* request
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum BlockResult {
+    V0(BlockResultV0),
+}
+
+/// Request latest events corresponding to the branch heads in a pub/sub topic
+///
+/// In response an Event is sent for each commit chunk that belong to branch heads
+/// that are not present in the requestor's known heads
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BranchHeadsReqV0 {
+    /// Topic public key of the branch
+    pub topic: TopicId,
+
+    /// Known heads
+    pub known_heads: Vec<ObjectId>,
+}
+
+/// Request latest events corresponding to the branch heads in a pub/sub topic
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum BranchHeadsReq {
+    V0(BranchHeadsReqV0),
+}
+
+/// Branch synchronization request
+///
+/// In response a stream of `Block`s of the requested Objects are sent
+/// that are not present in the requestor's known heads and commits
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct BranchSyncReqV0 {
+    /// Heads to request, including all their dependencies
+    pub heads: Vec<ObjectId>,
+
+    /// Fully synchronized until these commits
+    pub known_heads: Vec<ObjectId>,
+
+    /// Known commit IDs since known_heads
+    pub known_commits: BloomFilter,
+}
+
+/// Branch synchronization request
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum BranchSyncReq {
+    V0(BranchSyncReqV0),
+}
+
+/// Events the requestor needs, see EventReqV0
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct NeedEventsV0 {
+    /// Publisher ID
+    pub publisher: Digest,
+
+    /// First sequence number to request
+    pub from: u32,
+
+    /// Last sequence number to request
+    pub to: u32,
+}
+
+/// Events the responder has, see EventRespV0
+#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
+pub struct HaveEventsV0 {
+    /// Publisher ID
+    pub publisher: Digest,
+
+    /// First sequence number to send
+    pub from: u32,
+
+    /// Last sequence number to send
+    pub to: u32,
+}
+
+/// Request missed events for a pub/sub topic
+/// for the specified range of publisher sequence numbers
+///
+/// In response an EventResp then a stream of Events are sent
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct EventReqV0 {
+    /// Topic public key
+    pub topic: TopicId,
+
+    /// Events needed by the requestor
+    pub need: Vec<NeedEventsV0>,
+}
+
+/// Request missed events for a pub/sub topic
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum EventReq {
+    V0(EventReqV0),
+}
+
+/// Response to an EventReq
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct EventRespV0 {
+    /// Events the responder has
+    pub have: Vec<HaveEventsV0>,
+}
+
+/// Response to an EventReq
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum EventResp {
+    V0(EventRespV0),
+}
+
+/// Content of OverlayRequestV0
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayRequestContentV0 {
+    EventReq(EventReq),
+    BranchHeadsReq(BranchHeadsReq),
+    BranchSyncReq(BranchSyncReq),
+}
+
+/// Request sent to an overlay
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct OverlayRequestV0 {
+    /// Request ID
+    pub id: u64,
+
+    /// Request content
+    pub content: OverlayRequestContentV0,
+}
+
+/// Request sent to an overlay
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayRequest {
+    V0(OverlayRequestV0),
+}
+
+/// Content of OverlayResponseV0
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayResponseContentV0 {
+    Block(Block),
+    EventResp(EventResp),
+    Event(Event),
+}
+
+/// Request sent to an overlay
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct OverlayResponseV0 {
+    /// Request ID
+    pub id: u64,
+
+    /// Result
+    pub result: u8,
+
+    /// Response content
+    pub content: Option<OverlayResponseContentV0>,
+}
+
+/// Request sent to an OverlayRequest
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayResponse {
+    V0(OverlayResponseV0),
+}
+
+/// Content of PeerAdvertV0
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct PeerAdvertContentV0 {
+    /// Peer ID
+    pub peer: PeerId,
+
+    /// Topic subscriptions
+    pub subs: BloomFilter128,
+
+    /// Network addresses
+    pub address: Vec<NetAddr>,
+
+    /// Version number
+    pub version: u16,
+
+    /// App-specific metadata (profile, cryptographic material, etc)
+    #[serde(with = "serde_bytes")]
+    pub metadata: Vec<u8>,
+}
+
+/// Peer advertisement
+///
+/// Sent periodically across the overlay along random walks.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct PeerAdvertV0 {
+    /// Peer advertisement content
+    pub content: PeerAdvertContentV0,
+
+    /// Signature over content by peer's private key
+    pub sig: Sig,
+
+    /// Time-to-live, decremented at each hop
+    pub ttl: u8,
+}
+
+/// Peer advertisement
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum PeerAdvert {
+    V0(PeerAdvertV0),
+}
+
+/// Content of OverlayMessagePaddedV0
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayMessageContentV0 {
+    OverlayConnect(OverlayConnect),
+    OverlayDisconnect(OverlayDisconnect),
+    PeerAdvert(PeerAdvert),
+    TopicAdvert(TopicAdvert),
+    SubReq(SubReq),
+    SubAck(SubAck),
+    UnsubReq(UnsubReq),
+    UnsubAck(UnsubAck),
+    Event(Event),
+    ObjectSearchTopic(BlockSearchTopic),
+    ObjectSearchRandom(BlockSearchRandom),
+    ObjectResult(BlockResult),
+    OverlayRequest(OverlayRequest),
+    OverlayResponse(OverlayResponse),
+}
+
+/// Padded content of OverlayMessageV0
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct OverlayMessageContentPaddedV0 {
+    pub content: OverlayMessageContentV0,
+
+    /// Optional padding
+    #[serde(with = "serde_bytes")]
+    pub padding: Vec<u8>,
+}
+
+/// Overlay message
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct OverlayMessageV0 {
+    /// Overlay ID
+    pub overlay: OverlayId,
+
+    /// Session ID
+    pub session: SessionId,
+
+    /// Padded content encrypted with ChaCha20
+    /// - overlay_secret: BLAKE3 derive_key ("LoFiRe Overlay BLAKE3 key",
+    ///                                      repo_pubkey + repo_secret)
+    /// - key: BLAKE3 derive_key ("LoFiRe OverlayMessage ChaCha20 key",
+    ///                           overlay_secret + session_id)
+    /// - nonce: per-session message sequence number of sending peer
+    pub content: OverlayMessageContentPaddedV0,
+
+    /// BLAKE3 MAC
+    /// BLAKE3 keyed hash over the encrypted content
+    /// - key:  BLAKE3 derive_key ("LoFiRe OverlayMessage BLAKE3 key",
+    ///                            overlay_secret + session_id)
+    pub mac: Digest,
+}
+
+/// Overlay message
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum OverlayMessage {
+    V0(OverlayMessageV0),
+}
 
 //
 // BROKER PROTOCOL
@@ -98,8 +649,7 @@ pub enum DelClient {
 
 /// Content of `BrokerRequestV0`
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum BrokerRequestContentV0 {
-}
+pub enum BrokerRequestContentV0 {}
 /// Broker request
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct BrokerRequestV0 {
@@ -564,7 +1114,7 @@ pub enum AuthResult {
 }
 
 //
-// DIRECT MESSAGES
+// DIRECT / OUT-OF-BAND MESSAGES
 //
 
 /// Link/invitation to the repository
@@ -586,25 +1136,6 @@ pub enum RepoLink {
     V0(RepoLinkV0),
 }
 
-/// Owned repository with private key
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct RepoKeysV0 {
-    /// Repository private key
-    pub key: PrivKey,
-
-    /// Repository secret
-    pub secret: SymKey,
-
-    /// Peers to connect to
-    pub peers: Vec<PeerAdvert>,
-}
-
-/// Owned repository with private key
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum RepoKeys {
-    V0(RepoKeysV0),
-}
-
 /// Link to object(s) or to a branch from a repository
 /// that can be shared to non-members
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -623,80 +1154,21 @@ pub enum ObjectLink {
     V0(ObjectLinkV0),
 }
 
-//
-// BROKER STORAGE
-//
-
-/// A topic this node subscribed to in an overlay
+/// Owned repository with private key
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct TopicV0 {
-    /// Topic public key ID
-    pub id: PubKey,
+pub struct RepoKeysV0 {
+    /// Repository private key
+    pub key: PrivKey,
 
-    /// Signed `TopicAdvert` for publishers
-    pub advert: Option<TopicAdvert>,
-
-    /// Set of branch heads
-    pub heads: Vec<ObjectId>,
-
-    /// Number of local users that subscribed to the topic
-    pub users: u32,
-}
-
-/// A topic this node subscribed to in an overlay
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum Topic {
-    V0(TopicV0),
-}
-
-/// An overlay this node joined
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct OverlayV0 {
-    /// Overlay ID
-    pub id: OverlayId,
-
-    /// Overlay secret
+    /// Repository secret
     pub secret: SymKey,
 
-    /// Known peers with connected flag
+    /// Peers to connect to
     pub peers: Vec<PeerAdvert>,
-
-    /// Topics this node subscribed to in the overlay
-    pub topics: Vec<TopicId>,
-
-    /// Number of local users that joined the overlay
-    pub users: u32,
-
-    /// Last access by any user
-    pub last_access: Timestamp,
 }
 
-/// An overlay this node joined
+/// Owned repository with private key
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum Overlay {
-    V0(OverlayV0),
-}
-
-/// User account
-///
-/// Stored as user_pubkey -> Account
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct AccountV0 {
-    /// Authorized client pub keys
-    pub authorized_keys: Vec<PubKey>,
-
-    /// Admins can add/remove user accounts
-    pub admin: bool,
-
-    /// Overlays joined
-    pub overlays: Vec<OverlayId>,
-
-    /// Topics joined, with publisher flag
-    pub topics: Vec<TopicId>,
-}
-
-/// User account
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum Account {
-    V0(AccountV0),
+pub enum RepoKeys {
+    V0(RepoKeysV0),
 }
