+++ lofire-p2p/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod types;
+
